+++ packages/typespec-autorest/src/emit.ts
@@ -96,14 +96,6 @@ export function resolveAutorestOptions(
       NoTarget,
     );
   }
-  // eslint-disable-next-line @typescript-eslint/no-deprecated
-  if (resolvedOptions["arm-resource-flattening"]) {
-    reportDeprecated(
-      program,
-      `arm-resource-flattening option is deprecated and will be removed in future versions. Instead use \`@@Azure.ClientGenerator.Core.Legacy.flattenProperty(<ResourceName>.properties, "autorest");\``,
-      NoTarget,
-    );
-  }
 
   return {
     outputFile: resolvedOptions["output-file"],
@@ -119,8 +111,6 @@ export function resolveAutorestOptions(
     armTypesDir,
     useReadOnlyStatusSchema: resolvedOptions["use-read-only-status-schema"],
     emitLroOptions: resolvedOptions["emit-lro-options"],
-    // eslint-disable-next-line @typescript-eslint/no-deprecated
-    armResourceFlattening: resolvedOptions["arm-resource-flattening"],
     emitCommonTypesSchema: resolvedOptions["emit-common-types-schema"],
     xmlStrategy: resolvedOptions["xml-strategy"],
   };